requirements-cloud.txt
# Autoencoders Demo - Requirements for Streamlit Cloud
# Optimized for 1GB RAM limit, CPU-only deployment

# Deep Learning Framework (CPU version for smaller size)
tensorflow-cpu>=2.13.0,<2.16.0
keras>=2.13.0

# Data Science (specific versions for compatibility)
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scikit-learn>=1.3.0,<2.0.0
scipy>=1.11.0,<2.0.0

# Visualization
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0
plotly>=5.14.0,<6.0.0

# Image Processing
Pillow>=10.0.0,<11.0.0
scikit-image>=0.21.0,<1.0.0

# Dimensionality Reduction
umap-learn>=0.5.3,<1.0.0

# Web App
streamlit>=1.28.0,<2.0.0

# Utilities
tqdm>=4.66.0,<5.0.0
